refactor(fs): extract trash restore module
